requirements.txt
asttokens==2.2.1
audioread==3.0.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
debugpy==1.6.6
decorator==4.4.2
executing==1.2.0
Flask==2.2.3
gTTS==2.3.1
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
ipykernel==6.21.3
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.3.0
librosa==0.8.0
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
moviepy==1.0.3
nest-asyncio==1.5.6
numba==0.56.4
numpy==1.22.4
opencv-contrib-python==4.6.0.66
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pooch==1.7.0
proglog==0.1.10
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydub==0.25.1
pygame==2.3.0
Pygments==2.14.0
python-dateutil==2.8.2
pyzmq==25.0.1
requests==2.28.2
resampy==0.4.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
stack-data==0.6.2
threadpoolctl==3.1.0
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.63.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
